SmartHomeAssistant/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
curl_cffi==0.6.2
dataclasses-json==0.6.4
decorator==5.1.1
diskcache==5.6.3
duckduckgo_search==5.1.0
executing==2.0.1
fastapi==0.110.0
filelock==3.13.1
frozenlist==1.4.1
fsspec==2024.2.0
future==1.0.0
geocoder==1.38.1
google-ai-generativelanguage==0.4.0
google-api-core==2.17.0
google-auth==2.27.0
google-cloud-aiplatform==1.41.0
google-cloud-bigquery==3.17.2
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-generativeai==0.3.2
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.60.1
grpcio-status==1.60.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.4
httpx==0.27.0
hyperframe==6.0.1
idna==3.6
jedi==0.19.1
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.12
langchain-community==0.0.28
langchain-core==0.1.32
langchain-text-splitters==0.0.1
langsmith==0.1.27
llama_cpp_python==0.2.57
llvmlite==0.42.0
lxml==5.1.0
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.2.1
numba==0.59.0
numpy==1.26.4
openai-whisper==20231117
orjson==3.9.15
packaging==23.2
parso==0.8.3
pexpect==4.9.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
pycparser==2.21
pycryptodomex==3.20.0
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.1
ratelim==0.1.6
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rsa==4.9
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
soupsieve==2.5
SpeechRecognition==3.10.1
SQLAlchemy==2.0.28
sse-starlette==2.0.0
stack-data==0.6.3
starlette==0.36.3
starlette-context==0.3.6
sympy==1.12
tenacity==8.2.3
tiktoken==0.6.0
torch==2.2.1
tqdm==4.66.2
traitlets==5.14.1
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.28.0
vertexai==0.0.1
wcwidth==0.2.13
yarl==1.9.4

websockets~=12.0
pocketsphinx~=5.0.3